requirements.txt
asgiref==3.4.1
astroid==2.5.2
boto3==1.17.104
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
django==3.2.5
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.7.2
pytz==2021.1
python-dotenv==1.0.1
requests==2.25.1
rich==10.4.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
